requirements.txt
altgraph==0.17.4
Brotli==1.1.0
cffi==1.17.1
comtypes==1.4.11
dearpygui==2.0.0
Glymur==0.14.3
imageio==2.37.0
inflate64==1.0.3
lazy_loader==0.4
lxml==6.0.0
multivolumefile==0.2.3
networkx==3.5
numpy==2.3.1
opencv-python==4.11.0.86
packaging==25.0
# patoolib
pefile==2023.2.7
pillow==11.3.0
psutil==7.0.0
py7zr==1.0.0
pybcj==1.0.6
pycparser==2.22
pycryptodomex==3.23.0
pyglet==2.1.6
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
PyOpenGL==3.1.9
PyOpenGL-accelerate
pyppmd==1.2.0
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
pywin32-ctypes==0.2.3
pyzstd==0.17.0
scikit-image==0.25.2
scipy==1.16.0
setuptools==80.9.0
shiboken6==6.9.1
texttable==1.7.0
tifffile==2025.6.11
unrar2-cffi==0.4.1
pytest
pytest-qt
pytest-mock
